+++ R/jit-registry.R
@@ -77,6 +77,6 @@
   list(name = "nv_runif", static = c(1L, 3L, 4L, 5L)),
   list(name = "nv_rnorm", static = c(1L, 3L)),
   list(name = "nv_rbinom", static = c(1L, 3L, 4L, 5L)),
-  list(name = "nv_sample_int", static = c(1L, 3L, 4L)),
-  list(name = "nv_sample", static = 1L)
+  list(name = "nv_sample_int", static = c(1L, 3L, 4L, 6L)),
+  list(name = "nv_sample", static = c(1L, 4L))
 )